@@ -1,56 +1,66 @@
 {
-  "name": "angular-tour-of-heroes",
-  "version": "1.0.0",
-  "license": "MIT",
-  "scripts": {
-    "ng": "ng",
-    "start": "ng serve",
-    "build": "ng build",
-    "test": "ng test",
-    "lint": "ng lint",
-    "e2e": "ng e2e"
-  },
-  "private": true,
-  "dependencies": {
-    "@angular/animations": "6.0.0",
-    "@angular/cdk": "^6.0.0",
-    "@angular/common": "6.0.0",
-    "@angular/compiler": "6.0.0",
-    "@angular/core": "6.0.0",
-    "@angular/forms": "6.0.0",
-    "@angular/http": "6.0.0",
-    "@angular/material": "^6.0.0",
-    "@angular/platform-browser": "6.0.0",
-    "@angular/platform-browser-dynamic": "6.0.0",
-    "@angular/router": "6.0.0",
-    "angular-in-memory-web-api": "^0.6.0",
-    "body-parser": "^1.18.2",
-    "core-js": "^2.5.4",
-    "hammerjs": "^2.0.8",
-    "rxjs": "^6.1.0",
-    "zone.js": "^0.8.26"
-  },
-  "devDependencies": {
-    "@angular/cli": "6.0.0",
-    "@angular/compiler-cli": "6.0.0",
-    "@angular/language-service": "6.0.0",
-    "@types/jasmine": "~2.8.3",
-    "@types/jasminewd2": "~2.0.2",
-    "@types/node": "^6.0.103",
-    "codelyzer": "^4.2.1",
-    "concurrently": "^3.5.1",
-    "jasmine-core": "~2.8.0",
-    "jasmine-spec-reporter": "~4.2.1",
-    "karma": "~2.0.0",
-    "karma-chrome-launcher": "~2.2.0",
-    "karma-coverage-istanbul-reporter": "^1.4.2",
-    "karma-jasmine": "~1.1.0",
-    "karma-jasmine-html-reporter": "^0.2.2",
-    "protractor": "~5.1.2",
-    "ts-node": "~4.1.0",
-    "tslint": "~5.9.1",
-    "typescript": "2.7.2",
-    "webpack-bundle-analyzer": "^2.11.1",
-    "@angular-devkit/build-angular": "~0.6.0"
-  }
+    "name": "angular-tour-of-heroes",
+    "version": "1.0.0",
+    "license": "MIT",
+    "scripts": {
+        "ng": "ng",
+        "start": "ng serve",
+        "build": "ng build",
+        "test": "ng test",
+        "lint": "ng lint",
+        "e2e": "ng e2e"
+    },
+    "private": true,
+    "dependencies": {
+        "@angular/animations": "6.1.2",
+        "@angular/cdk": "^6.4.3",
+        "@angular/common": "6.1.2",
+        "@angular/compiler": "6.1.2",
+        "@angular/core": "6.1.2",
+        "@angular/forms": "6.1.2",
+        "@angular/http": "6.1.2",
+        "@angular/material": "^6.4.3",
+        "@angular/platform-browser": "6.1.2",
+        "@angular/platform-browser-dynamic": "6.1.2",
+        "@angular/router": "6.1.2",
+        "@ngrx/effects": "6.1.0",
+        "@ngrx/entity": "6.1.0",
+        "@ngrx/router-store": "6.1.0",
+        "@ngrx/store": "6.1.0",
+        "@ngrx/store-devtools": "^6.1.0",
+        "angular-in-memory-web-api": "^0.6.0",
+        "body-parser": "^1.18.2",
+        "core-js": "^2.5.4",
+        "hammerjs": "^2.0.8",
+        "rxjs": "^6.2.2",
+        "zone.js": "^0.8.26"
+    },
+    "devDependencies": {
+        "@angular-devkit/build-angular": "~0.6.0",
+        "@angular-devkit/build-ng-packagr": "~0.7.0",
+        "@angular/cli": "6.1.3",
+        "@angular/compiler-cli": "^6.0.5",
+        "@angular/language-service": "6.1.2",
+        "@ngrx/schematics": "^5.2.0",
+        "@types/jasmine": "~2.8.3",
+        "@types/jasminewd2": "~2.0.2",
+        "@types/node": "^6.0.103",
+        "codelyzer": "^4.2.1",
+        "concurrently": "^3.5.1",
+        "jasmine-core": "~2.8.0",
+        "jasmine-spec-reporter": "~4.2.1",
+        "karma": "~2.0.0",
+        "karma-chrome-launcher": "~2.2.0",
+        "karma-coverage-istanbul-reporter": "^1.4.2",
+        "karma-jasmine": "~1.1.0",
+        "karma-jasmine-html-reporter": "^0.2.2",
+        "ng-packagr": "^3.0.0",
+        "protractor": "~5.1.2",
+        "ts-node": "~4.1.0",
+        "tsickle": ">=0.29.0",
+        "tslib": "^1.9.0",
+        "tslint": "~5.9.1",
+        "typescript": "^2.7.2",
+        "webpack-bundle-analyzer": "^2.11.1"
+    }
 }
